[BE] feat: (49) Topic Controller 수정

+++ BE/colors/src/main/java/com/ssafy/colors/response/TopicDTO.java
@@ -9,7 +9,7 @@ public class TopicDTO {
     private Long id;
     private String title;
     private boolean isRecommend;
-    private int cnt;
+    private long cnt;
